+++ src/main/java/org/folio/linked/data/model/entity/Resource.java
@@ -8,7 +8,6 @@ import static java.util.Objects.isNull;
 import static java.util.Objects.nonNull;
 import static java.util.Optional.ofNullable;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.INSTANCE;
-import static org.springframework.util.CollectionUtils.isEmpty;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import io.hypersistence.utils.hibernate.type.json.JsonBinaryType;
@@ -45,6 +44,7 @@ import lombok.experimental.Accessors;
 import org.folio.ld.dictionary.ResourceTypeDictionary;
 import org.folio.linked.data.configuration.audit.LinkedDataAuditEntityListener;
 import org.folio.linked.data.validation.PrimaryTitleConstraint;
+import org.folio.linked.data.validation.ResourceTypeConstraint;
 import org.folio.marc4ld.util.ResourceKind;
 import org.hibernate.annotations.Type;
 import org.hibernate.annotations.UpdateTimestamp;
@@ -56,6 +56,7 @@ import org.springframework.data.domain.Persistable;
 @Data
 @NoArgsConstructor
 @PrimaryTitleConstraint
+@ResourceTypeConstraint
 @Accessors(chain = true)
 @Table(name = "resources")
 @EqualsAndHashCode(of = "id")
@@ -220,9 +221,6 @@ public class Resource implements Persistable<Long> {
   @PrePersist
   void prePersist() {
     this.managed = true;
-    if (isEmpty(types)) {
-      throw new IllegalStateException("Cannot save resource [" + id + "] without types");
-    }
     if (nonNull(folioMetadata) && !isOfType(INSTANCE) && !isAuthority()) {
       throw new IllegalStateException("Cannot save resource [" + id + "] with types " + types + ". "
         + "Folio metadata can be set only for instance and authority resources");